refactor: add DistanceFare
