+++ src/main/java/nextstep/subway/domain/model/FareCalculator.java
@@ -1,0 +1,19 @@
+package nextstep.subway.domain.model;
+
+public class FareCalculator {
+    public static final int BASE_FARE = 1250;
+    public static final int EXTRA_FARE_PER_5KM = 100;
+    public static final int EXTRA_FARE_PER_8KM = 100;
+    public static final int DISTANCE_THRESHOLD_10KM = 10;
+    public static final int DISTANCE_THRESHOLD_50KM = 50;
+    public static final int UNIT_DISTANCE_5KM = 5;
+    public static final int UNIT_DISTANCE_8KM = 8;
+    public static final int MAX_EXTRA_FARE_10KM_TO_50KM =
+        (int) Math.ceil((DISTANCE_THRESHOLD_50KM - DISTANCE_THRESHOLD_10KM) / (double) UNIT_DISTANCE_5KM) * EXTRA_FARE_PER_5KM;
+
+
+    public static int calculateFare(int distance) {
+        DistanceFare distanceFare = DistanceFare.from(distance);
+        return distanceFare.calculateFare(distance);
+    }
+}
