+++ src/test/java/nextstep/subway/application/FareCalculatorTest.java
@@ -1,11 +1,13 @@
 package nextstep.subway.application;
 
-import static nextstep.subway.application.FareCalculator.*;
+import static nextstep.subway.domain.model.FareCalculator.*;
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import nextstep.subway.domain.model.FareCalculator;
 
 class FareCalculatorTest {
